// gen/ts-fetch/models/TenantUser.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TenantUser = {
    userId?: string;
    resellerId?: string | null;
    enterpriseId?: string | null;
    email?: string;
    displayName?: string;
    role?: string;
    status?: string;
    assignedEnterpriseIds?: Array<string> | null;
    departmentId?: string | null;
    createdAt?: string;
};